libdraw, devdraw: change representation of scroll events in mouse messages
Instead of having two mouse button events for scrolling one line at
a time, define a single type of scroll event (as a virtual button, if
you will) and extend mouse messages to carry signed scroll distance
(in lines).

This change lays the ground for future support of high-precision
scrolling which is impossible without passing exact scroll offsets
through to a client program. In addition, it is also useful even in
the realm of discrete line-granular scroll, since it will allow
platforms to generate a single message with a line offset, when it
is available, instead of spawning a stream of single-line scroll-up
or scroll-down messages.

Do not use any of the more advanced APIs for tracking scroll offsets
in the platform code yet; just provide the ±1 scroll distances when
triggering mouse tracking calls during scrolling.