devdraw: add support for precise pixel-based scrolling on macOS
Change generation of mouse messages for scroll events to implement
the mechanics of precise pixel-granular scrolling on macOS. All the
extra information necessary is already available in the NSEvent object
passed to the scrollWheel() method. This allows us to take advantage
of smooth scrolling regimes supported by high-precision input devices,
such as trackpads, Magic Mouse, mice with free-spinning scroll wheels,
trackpoints, etc.

The changes in libdraw to support scrolling in pixel increments are
minimal: Add a new type of scroll event (as another virtual button)
that has the same logic as the existing one for line-based scrolling,
but uses precise pixel offsets instead of coarse line offsets for
scroll distances. Additionally, add two more events for marking the
beginning and ending of series of these pixel-granular scroll events
that are generated by a single motion or gesture on a high-precision
input device. The latter two events can be used to chain the offsets
in such series to implement smooth line-snapping scrolling in client
programs (for example, it text frames).

Make sure not to signal a motion stop without a prior motion start,
which would be possible if we did not keep track ourselves, as an
NSEventPhaseCancelled may arrive following an NSEventPhaseMayBegin
without an NSEventPhaseBegan materializing. For a detailed discussion
of handling trackpad scrolling events, see Cocoa documentation at
<https://developer.apple.com/library/archive/documentation/Cocoa/Conceptual/EventOverview/HandlingTouchEvents/HandlingTouchEvents.html>.