acme: go display line's worth of runes for long lines in textbacknl()
Change the behavior of textbacknl() in regards to traversing long
lines. Previously, it would consider one line (out of n) to be passed
after either encountering a newline or scanning 128 runes without
running into one — for performance reasons. Thus, technically, it
counted not lines, but chunks of at most 128 non-newline runes.

Now, instead of using an arbitrary bound to break a long line, we
switch to tracking the total width of characters encountered since
we last declared a line break, and stop once the next rune is about
to push it over the width of the frame. Effectively, we switch to
counting display lines — if we presuppose that the starting position
p itself is at the start of a display line.

This assumption obviously holds when we call textbacknl() to scroll
the text in a frame (upward) a certain amount of lines (like in
textscroll() and textframescroll()). In fact, that use case is the
main reason for this change. With it, we lay the ground for more
consistent mouse scrolling experience. Using the old textbacknl(),
when we encounter a long line scrolling up, we might bring to the
screen more than one display line for a narrow frame or much less
than one display line’s worth of runes for a wide frame, resulting
in, respectively, faster or slower scrolling rates compared to those
for short lines fitting to the frame’s width. However, when scrolling
down, the same line will appear in chunks of full frame width, not
necessarily at the same rate. With the new textbacknl(), scrolling
becomes predictably linear in that the number of display lines going
past the viewport stays the same for the same input regardless
of physical line lengths or scrolling direction.

Granted, not all calls to textbacknl() pass it a starting position
that is necessarily at a display line boundary. However, in all of
those remaining calls, we do not need the resulting position to be
in accordance with the current cadence of display line boundaries.
Instead, we are looking backward a loosely defined number of lines
(0, 1, 2, ¼ or ¾ of viewport’s height) to give some visual context
to a certain place we would like to bring to the viewport, so going
back n display lines’ worth of runes can only be more pleasing
visually, compared to going back a strict 128*n runes.

Performance-wise this change is also fine. Thanks to font caching,
width-counting is not penalizing, really. As to the number of runes
being scanned, although for large screens and small fonts a display
line may fit more than 128 runes, its length is still going to be
capped by some reasonable bound, and using it as a measure is not
going to cost us exceedingly: If it is fine to scan as many runes
per display line to draw them, it should be fine to do the same
for scrolling.