acme: add textforwardnl(), forward-moving counterpart to textbacknl()
Like the newly updated textbacknl(), textforwardnl() breaks long lines
based on their display width, so that if the starting position p is at
the start of a display line itself, we skip exactly n display lines.

Together, these two functions provide the building blocks for moving
viewport an arbitrary amount of display lines backwards of forwards,
which is necessary for brining smooth inertial scrolling to acme (via
the new format of mouse messages for scrolling events that is now
supported by devdraw).