libdraw: support configurable scaling multipliers for mouse scrolling
Provide two separate functions for scaling scroll distances that are
in pixels and lines, respectively. Client programs are expected to
call them when handling mouse scroll events. Clients relying on the
mouselinesnapscroll() function to implement line-snapping scrolling
are relieved from that, as it makes the necessary calls internally.

The mousepixelscrollscaled() and mouselinescrollscaled() functions
are supposed to supersede mousescrollsize(), which should be removed
as soon as acme and 9term — the only two clients supporting mouse
scrolling — are transitioned to the new mouse scroll messaging.

Scale factors for pixel-based and line-based scrolling are configured
separately, due to the corresponding events typically originating from
input devices of different scrolling precision by nature (turns of
a notched mouse scroll wheel might come in at different rates than
a series of scrolls spawned by a flick of fingers on a trackpad).

For that, two new environment variables ‘mousepixelscrollscale’ and
‘mouselinescrollscale’ are recognized. Both expected to be set to
strictly positive floating-point numbers in the format understood by
strtod(). The latter variable allows the number to be followed by the
‘vh’ suffix of the same meaning as in CSS: that the value is given in
percent of viewport’s height. For backward compatibility, when
‘mouselinescrollscale’ is unset, ‘mousescrollsize’ is parsed for the
multiplier instead, following exactly the same format, except for the
viewport-relative suffix being ‘%’ and not ‘vh’.

Further, for line-based scrolling in particular, it is also allowed
to set the lowest number of lines to scroll at a time, which is
enforced after the scaling multiplier is applied. For that, a new
environment variable ‘mouselinescrollmin’ is recognized that is
expected to be set to a non-negative integer.

By default, both scaling multipliers are set to 1.0, and the minimum
line scroll amount is set to 1 line.