acme: update size of expanded tags on mouse-triggered cut and paste
Make sure that when cut() and paste() are called for the 1-2 and 1-3
mouse chords in a window's tag that is currently expanded, its size is
recomputed, and the window is updated for the new tag height. If the
tag is not expanded but has gotten some text pasted at the very end so
that the new dot is completely out of sight, expand the tag, too. This
brings it in line with cutting and pasting via the Cmd+X and Cmd+V
keyboard shortcuts which already do that (correctly).