acme: refactor nil checks and usage of window pointers in mouse handlers
Restructure the three branches handling events related to the mouse
crossing window boundaries. Remove superfluous checks for t!=nil and
m.buttons!=0. For clicks and drags on scrollbars and layout boxes,
take into account that neither a tag or a body can have its w be nil,
and that neither a tag or a column tag can have its col be nil.