devdraw: ensure imprecise scrolling deltas are at least ±1 on macOS
It is my understanding that for imprecise scrolling events (e.g.,
coming from a notched scroll wheel) scrollingDeltaY is supposed to
contain a nonzero integer value (in a CGFloat). It is said in the
AppKit documentation to multiply the delta by the line or row height
to get the scroll offset in pixels in that mode, but it is not clear
whether the delta may be fractional. If it may, it is possible to end
up with zero line scroll distance after truncation or rounding. Just
in case it happens, round tiny deltas up to at least ±1.