acme: ensure multiline tags are expanded after loading from a dump file
Since the state of windows’ tagexpand flags is not recorded in dump
files and new windows have it set to true by default, it makes sense
to conform with that default when restoring from a dump, and ensure
that tags are resized to fit their content after loading. Otherwise,
we may leave a tag in an inconsistent state where tagexpand is set to
true, but the tag is not expanded (which also means that the first
attempt to expand a tag manually by scrolling down will be ignored
until it is first “collapsed” by scrolling up).