acme: when typing in a tag, expand and scroll it if dot is out of sight
Currently, a window’s tag is expanded after a key press only when
a newline is entered. To avoid blind typing, do so also when the dot
is out of sight (once the key takes its effect). If the tag is already
set to expand, proceed to tag resizing right away. If the dot is still
hidden after that, scroll the tag (up or down) to bring it to screen.

If the key being pressed is bound to tag scrolling and therefore takes
its full effect in texttype(), skip tag resizing altogether. This way,
if the key just scrolled, collapsed, or expanded the tag, we do not
accidentally negate the user’s intention in case the dot was left in
the now hidden part. Additionally, if the key just expanded the tag,
we avoid a superfluous winresize() call.

This change brings the behavior of the tag in line with the behavior
of the body, which always scrolls to show the dot when typing if it
happens to be outside of the viewport.

While at it, move the invocation of tag resizing from rowtype() to
wintype() as the lowest place in the call tree that feels natural for
such update. This does not create any side effects to be wary of, as
the only place wintype() is called from is rowtype(), anyway.