acme: make tags scrollable when expanded
When scrolling down, a tag is expanded if it is collapsed; after that
it is scrolled the same way a body is, except the viewport is not
allowed to scroll past the last line with content. When expanded in
a mouse scroll motion, however, it does not scroll (by mouse) until
that motion ends so that when, say, a touchpad is used, the first
swipe expands a collapsed tag but does not go on to kick off scrolling
immediately; only the following swipe does that.

When scrolling up, a tag is scrolled the same way a body is until the
very top is reached; subsequent upward scroll events collapse it. When
brought to the top in a mouse scroll motion, though, it does not
collapse (by mouse scrolling) until that motion ends so that when
a touchpad is used, one can swipe to scroll up freely without fearing
to accidentally collapse the tag if that input is enough to hit the
top; one more swipe is necessary to trigger a collapse.