libmemdraw: use not ulong but u32int words for allocating image data
_allocmemimage() allocates image data in units of ulong words, while
the rest of the code (correctly) uses them as u32int words. Because
of this, on 64-bit platforms where sizeof(ulong)==8, Memdata storage
of each image is over-allocated by a factor of two! We fix this by
changing sizeof(ulong) to sizeof(u32int) in the computation of the
memory allocation size.

In order to store the Memdata back pointer and the caller PC right
next to the actual data without breaking strict aliasing (which modern
compilers are eager to take advantage of to the point of sometimes
optimizing violating code away), we use an intermediary Memalloc
struct explicitly containing those extra fields next to a flexible
array member for the u32int words of pixel data. We do not use the
array field beyond taking its address, but having it as a part of
the allocation struct guarantees correct alignment and safe access
free of undefined behavior.

Local variables holding image bit depth are renamed from ‘d’ to
‘depth’ in order to make it obvious at a glance that those are
unrelated to ‘md’, data size or anything like that. Finally, minor
inconsistencies in nil comparisons and code formatting are fixed.