libmemdraw: increment Memdata reference count in allocmemimaged()
When an existing Memdata is reused for a new Memimage via a call to
allocmemimaged(), currently it is a responsibility of the caller to
increment the Memdata ‘ref’ counter field. Make allocmemimaged() be
in charge of that instead. The transition to the new behavior is easy,
as at the moment there is only one place in the codebase (outside of
libmemdraw) where allocmemimaged() is called at all — memlalloc().