acme: replace \n with spaces when pasting into row or column tags
It is clumsy enough to undo pasting a long string into a column tag or
the row tag, but it gets even clumsier if the string happens to have
newline characters in it. Let us punish the user a bit less for such
mistake and replace newlines with spaces. Besides, texttype() already
prohibits typing the newline character into both row and column tags.