libcomplete: put output in one allocation; comply with strict aliasing
Make the complete() interface similar to the dirreadall() one in the
sense that all the returned data is contained in a single allocation
that can be released with a single free() call.

Further, eliminate unnecessary auxiliary allocations for matching file
names and filter the entries directly in the Dir[] array returned by
dirreadall(); this way, the filtering code simplifies, as well.

Finally, fix pointer-casting type punning to comply with modern strict
aliasing rules by only treating the allocated memory as a char array
and writing the Completion struct and the pointers in its filename[]
array via memcpy() (the calls to it are optimized away).