acme: fix so that in a tag's left margin, only layout box column-drags
Since a tag's left margin (the interior of its t->scrollr) below the
layout box is not visually separated from the rest of the tag, it is
easy to accidentally click on it while trying to interact with the
text of the tag without realizing it. Hence, process such clicks as
regular text-manipulating ones.